add divnonan
