+++ tfjs-core/src/ops/squared_difference.ts
@@ -27,7 +27,6 @@ import {assertAndGetBroadcastShape} from './broadcast_util';
 import {op} from './operation';
 import {scalar} from './tensor_ops';
 
-
 /**
  * Returns (a - b) * (a - b) element-wise.
  * Supports broadcasting.
